pyproject.toml
[project]
name = "mvpv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "nltk>=3.9.3",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "uvicorn>=0.41.0",
]
